Conflicts:
	src/corelib/global/global.pri
	src/corelib/global/qcompilerdetection.h
	src/corelib/global/qglobal.h
	src/corelib/tools/qdatetime.cpp
	src/plugins/platforms/xcb/qxcbscreen.h
	src/plugins/platforms/xcb/qxcbwindow.h
	src/widgets/dialogs/qcolordialog.cpp
	src/widgets/dialogs/qcolordialog_p.h
	tools/configure/configureapp.cpp


cregit: b6191b16d41459ed73cea738dfaf8e25e81ae22b https://github.com/qt/qtbase/commit/b6191b16d41459ed73cea738dfaf8e25e81ae22b

+++ src/corelib/tools/qbytearraymatcher.cpp
@@ -947,7 +947,7 @@ parameter_list|(
 name|a
 parameter_list|)
 define|\
-value|if (sl_minus_1< sizeof(uint) * CHAR_BIT) \         hashHaystack -= (a)<< sl_minus_1; \     hashHaystack<<= 1
+value|if (sl_minus_1< sizeof(uint) * CHAR_BIT) \         hashHaystack -= uint(a)<< sl_minus_1; \     hashHaystack<<= 1
 end_define
 begin_comment
 comment|/*!     \internal  */
